+++ src/enzo/ConstructFeedbackZone.C
@@ -36,7 +36,7 @@ int CommunicationReceiveHandler(fluxes **SubgridFluxesEstimate[] = NULL,
 
 
 
-grid* ConstructFeedbackZone(ActiveParticleType* ThisParticle,int FeedbackRadius,
+grid* ConstructFeedbackZone(ActiveParticleType* ThisParticle, FLOAT FeedbackRadius,
 			    FLOAT dx, HierarchyEntry** Grids, int NumberOfGrids,
 			    int SendField)
 {
